Fixed the fork bomb bug
If fork fails, wait. Also, maintain an "available processes" counter.
When it hits zero, wait for processes to finish. There is still a bug
where the "available processes" variable does not restrict the number of
processes that are actually created (this is only happening on my
machine, it works fine on SEAS). Seems to be that numAvailableProcesses
is never hitting zero so never triggers a wait....